data-engineering/coingecko_api/requirements.txt
# python library dependencies
python-dotenv
pycoingecko #coingecko API wrapper
sqlalchemy
psycopg2
pandas
mypy
